Move mailbox interface and factories to separate files
